+++ src/mpservice/_streamer.py
@@ -38,6 +38,7 @@ import random
 import threading
 import time
 import traceback
+from abc import ABC, abstractmethod
 from collections import deque
 from collections.abc import AsyncIterable, AsyncIterator, Iterable, Iterator, Sequence
 from inspect import iscoroutinefunction
@@ -57,6 +58,7 @@ import asyncstdlib.itertools
 from typing_extensions import Self  # In 3.11, import this from `typing`
 
 import mpservice.multiprocessing
+import mpservice.multiprocessing.remote_exception
 import mpservice.threading
 
 from ._common import StopRequested
@@ -2274,20 +2276,24 @@ class IterableQueue(Iterator[T]):
         self._q = q
         self._to_stop = to_stop
         self._wait_interval = 1.0
+        # User may revise this value after the object is initiated.
 
         self._num_suppliers = num_suppliers
         if isinstance(q, (queue.Queue, queue.SimpleQueue)):
-            self._spare_lids = queue.Queue(maxsize=0)
+            self._spare_lids = queue.Queue(maxsize=num_suppliers)
             self._applied_lids = queue.Queue(maxsize=num_suppliers)
-            self._removed_lids = queue.Queue(maxsize=num_suppliers)
+            self._used_lids = queue.Queue(maxsize=num_suppliers)
+            # self._lock = threading.Lock()
         else:
-            self._spare_lids = mpservice.multiprocessing.Queue(maxsize=0)
+            self._spare_lids = mpservice.multiprocessing.Queue(maxsize=num_suppliers)
             self._applied_lids = mpservice.multiprocessing.Queue(maxsize=num_suppliers)
-            self._removed_lids = mpservice.multiprocessing.Queue(maxsize=num_suppliers)
+            self._used_lids = mpservice.multiprocessing.Queue(maxsize=num_suppliers)
+            # self._lock = mpservice.multiprocessing.Lock()
         for _ in range(num_suppliers):
             self._spare_lids.put(None)
-        self._spare_lids.put('')
         # User should not touch these internal helper queues.
+        # TODO: the name 'lid' is not very good; something implying the "bottom" would be better.
+        # TODO: do we need to use a lock to group the access to the helper queues?
 
     def __getstate__(self):
         # This will fail if the queues are not pickle-able. That would be a user mistake.
@@ -2298,7 +2304,7 @@ class IterableQueue(Iterator[T]):
             self._num_suppliers,
             self._spare_lids,
             self._applied_lids,
-            self._removed_lids,
+            self._used_lids,
             self._can_timeout,
         )
 
@@ -2310,7 +2316,7 @@ class IterableQueue(Iterator[T]):
             self._num_suppliers,
             self._spare_lids,
             self._applied_lids,
-            self._removed_lids,
+            self._used_lids,
             self._can_timeout,
         ) = zz
 
@@ -2320,6 +2326,7 @@ class IterableQueue(Iterator[T]):
             return self._q.maxsize
         except AttributeError:
             return self._q._maxsize
+        # If you used a SimpleQueue for `__init__`, this would raise `AttributeError`.
 
     def qsize(self) -> int:
         return self._q.qsize()
@@ -2372,35 +2379,71 @@ class IterableQueue(Iterator[T]):
                     raise StopRequested
                 time.sleep(0)  # force a context switch
 
-    def put_end(self) -> None:
+    def put_end(self, *, wait_for_renew: bool = False) -> None:
         """
         Each "supplier" must call this method exactly once, after it is done putting
         data in the queue. Do not use `put(None)` for this purpose.
+
+        Suppose in a certain use case a queue is populated by a supplier, which has called `put_end`;
+        a consumer is designed to call `renew` after it finishes iterating the queue, allowing
+        the next round of data population/consumption. Before the consumer calls `renew`,
+        this object does not forbid the supplier from calling `put` to add new data elements
+        to the queue (unless the user imposes such restriction themselves), although these
+        new data elements are not accessible via `__next__` or `__iter__` until the consumer
+        has called `renew`. Suppose the supplier gets to call `put_end` before the consumer
+        calls `renew`, the object is in an unexpected state. If `wait_for_renew` is `True`,
+        this situation is allowed, and `put_end` will wait to go through once `renew` is called.
+        If `wait_for_renew` is `False` (the default), exception is raised in this situation.
         """
-        z = self._spare_lids.get()
-        if z == '':
-            raise RuntimeError('`put_end()` has been called too many times')
-        self._applied_lids.put(None)
+        # TODO: add an overall `timeout`?
+        if wait_for_renew:
+            while True:
+                try:
+                    z = self._spare_lids.get(timeout=self._wait_interval)
+                    break
+                except queue.Empty:
+                    if self._to_stop is not None and self._to_stop.is_set():
+                        raise StopRequested
+        else:
+            try:
+                z = self._spare_lids.get(timeout=0.01)
+            except queue.Empty:
+                raise RuntimeError(
+                    '`put_end` is called more than `num_suppliers` times'
+                )
+
+        self._applied_lids.put(z)
         self.put(None)
         # A `None` in the queue corresponds to a `None` in `self._applied_lids`.
 
     def __next__(self) -> T:
+        if self._used_lids.full():
+            raise StopIteration
+
         z = self.get()
         if z is None:
-            if self._removed_lids.full():
+            if self._used_lids.full():
                 # Other consumers have removed all the lids and confirmed
                 # there's no more data to come from the queue.
                 # There's no more `None` in `self._applied_lids`.
                 self.put(None)
                 # Let there always be an end marker so that other consumers
                 # can still iterate over this queue and see it's finished.
-                # `self._removed_lids` remains full, hence the next call
+                # `self._used_lids` remains full, hence the next call
                 # to `__next__` will get here again.
+                # This does not increase the number of `None`s in the queue
+                # as it simply replaces the one that is just taken off the queue.
                 raise StopIteration
-            self._applied_lids.get()
-            self._removed_lids.put(None)
-            if self._removed_lids.full():
+            z = self._applied_lids.get()
+            self._used_lids.put(z)
+            if self._used_lids.full():
                 # This is the first consumer who sees the queue is exhausted.
+                # Put an extra `None` in the queue for other consumers to see.
+                # This is needed because we don't assume nor limit the number
+                # of consumers to the queue.
+                # This is the only extra `None`: there is only one consumer
+                # who is the first to see the bottom of the queue, and subsequent
+                # consumers will get/put this `None` without increasing its count.
                 self.put(None)
                 raise StopIteration
             # The queue is not exhausted because all suppliers's end markers ("lids")
@@ -2414,3 +2457,133 @@ class IterableQueue(Iterator[T]):
                 yield self.__next__()
             except StopIteration:
                 break
+
+    def renew(self):
+        """
+        This is for special use cases where the queue needs to be "reused" for
+        more than one round of iterations.
+        In those use cases, typically the consumer (or consuming side if there are
+        more than one consumers) calls `renew` exactly once upon finishing iteration
+        over the content of the queue. The suppliers can put more data into the queue
+        and call `put_end` as usual once done; the consumer then iterates over the queue
+        as if there were no previous rounds.
+
+        This method can only be called after one round of consumption (by iteration) is complete.
+        It can not be called at the beginning when no data has been placed in the queue,
+        because the implementation does not provide a way to tell the object is in such "brand new"
+        state.
+
+        The application needs to ensure `renew` is called only once after one round of iteration.
+        """
+        if not self._used_lids.full():
+            raise RuntimeError('the object is not in a renewable state')
+        z = self.get()  # take out the extra `None`
+        if z is not None:
+            raise RuntimeError(f'expecting None, got {z}')
+
+        for _ in range(self._num_suppliers):
+            z = self._used_lids.get()
+            self._spare_lids.put(z)
+
+
+class ProcessRunnee(ABC):
+    def __enter__(self):
+        return self
+
+    def __exit__(self, *args):
+        pass
+
+    @abstractmethod
+    def __call__(self, *args, **kwargs) -> Any:
+        raise NotImplementedError
+
+
+class ProcessRunner:
+    """
+    `ProcessRunner` creates a custom object (of a subclass of `ProcessRunnee`)
+    in "background" process, and calls the object's `__call__` method as needed,
+    any number of times, while keeping the background process alive until `join` is called.
+
+    Some initial motivations for this class include:
+
+    - The custom object may perform nontrivial setup once, and keep the setup in effect
+      through the object's lifetime.
+
+    - User may pass certain data to the custom object's `__init__` via `ProcessRunner.__init__`.
+      A particular use case is to pass in a `mpservice.multiprocessing.Queue` or `mpservice.streamer.IterableQueue`,
+      because these objects can't be passed to the custom object via `ProcessRunner.restart`.
+
+      Note that a `mpservice.multiprocessing.Queue` can't be contained in another `mpservice.multiprocessing.Queue`,
+      nor can it be passed to a `mpservice.multiprocessing.Manager` in a call to a "proxy method".
+      If either is possible, the class `ProcessRunner` is not that needed.
+
+    In some use cases, `ProcessRunner`, along with `IterableQueue` and its method `renew`, are useful in stream processing.
+    """
+
+    def __init__(
+        self,
+        *,
+        target: type[ProcessRunnee],
+        args=None,
+        kwargs=None,
+        name=None,
+    ):
+        self._instructions = mpservice.multiprocessing.Queue()
+        self._result = mpservice.multiprocessing.Queue(maxsize=1)
+        self._process = mpservice.multiprocessing.Process(
+            target=self._work,
+            kwargs={
+                'instructions': self._instructions,
+                'result': self._result,
+                'worker_cls': target,
+                'args': args or (),
+                'kwargs': kwargs or {},
+            },
+            name=name,
+        )
+
+    @staticmethod
+    def _work(
+        instructions,
+        result,
+        worker_cls: type[ProcessRunnee],
+        args,
+        kwargs,
+    ):
+        with worker_cls(*args, **kwargs) as worker:
+            while True:
+                zz = instructions.get()
+                if zz is None:
+                    break
+                try:
+                    z = worker(*zz[0], **zz[1])
+                except Exception as e:
+                    z = mpservice.multiprocessing.remote_exception(e)
+                result.put(z)
+
+    def start(self):
+        self._process.start()
+
+    def join(self, timeout=None):
+        self._instructions.put(None)
+        self._process.join(timeout=timeout)
+
+    def restart(self, *args, **kwargs):
+        self._instructions.put((args, kwargs))
+
+    def rejoin(self):
+        z = self._result.get()
+        if isinstance(z, Exception):
+            raise z
+        return z
+
+    def __enter__(self):
+        self.start()
+        return self
+
+    def __exit__(self, *args):
+        self.join()
+
+    def __call__(self, *args, **kwargs):
+        self.restart(*args, **kwargs)
+        return self.rejoin()